extract compare mode card to component; fixup
